Move the side effect out of ProjectsContext
The side effect in ProjectsContext causes errors in jest testing.
Move it out to a separate hook useAutoFetchPipelines and place it into
useEnsureValidPipeline
