+++ services/orchest-webserver/client/src/contexts/ProjectsContext.tsx
@@ -1,12 +1,20 @@
-import { useFetchPipelines } from "@/hooks/useFetchPipelines";
-import { useHasChanged } from "@/hooks/useHasChanged";
 import type { PipelineMetaData, Project } from "@/types";
 import React from "react";
-import { KeyedMutator } from "swr";
 
 const ProjectsContext = React.createContext<IProjectsContext>(
   {} as IProjectsContext
 );
+
+export type IProjectsContextState = {
+  projectUuid?: string;
+  pipelineIsReadOnly: boolean;
+  pipelineSaveStatus: "saved" | "saving";
+  pipelines: PipelineMetaData[] | undefined;
+  pipeline?: PipelineMetaData | undefined;
+  projects: Project[];
+  hasLoadedProjects: boolean;
+  hasLoadedPipelinesInPipelineEditor: boolean;
+};
 
 export const useProjectsContext = () => React.useContext(ProjectsContext);
 
@@ -49,21 +57,11 @@ type Action =
     };
 
 type ActionCallback = (currentState: IProjectsContextState) => Action;
-type ProjectsContextAction = Action | ActionCallback;
-export interface IProjectsContextState {
-  projectUuid?: string;
-  pipelineIsReadOnly: boolean;
-  pipelineSaveStatus: "saved" | "saving";
-  pipelines: PipelineMetaData[] | undefined;
-  pipeline?: PipelineMetaData | undefined;
-  projects: Project[];
-  hasLoadedProjects: boolean;
-  hasLoadedPipelinesInPipelineEditor: boolean;
-}
+export type ProjectsContextAction = Action | ActionCallback;
+
 export interface IProjectsContext {
   state: IProjectsContextState;
   dispatch: (value: ProjectsContextAction) => void;
-  fetchPipelines: KeyedMutator<PipelineMetaData[]>;
 }
 
 const reducer = (
@@ -169,45 +167,8 @@ const initialState: IProjectsContextState = {
 export const ProjectsContextProvider: React.FC = ({ children }) => {
   const [state, dispatch] = React.useReducer(reducer, initialState);
 
-  const {
-    pipelines,
-    isFetchingPipelines,
-    error,
-    fetchPipelines,
-  } = useFetchPipelines(state.projectUuid);
-
-  const hasPipelinesChanged = useHasChanged(state.pipelines);
-
-  React.useEffect(() => {
-    // When switching projects, state.pipelines will be cleaned up.
-    // then we need to refetch.
-    // Note that we don't want to trigger this when page is just loaded.
-    // because useFetchPipelines will do the inital request.
-    const hasPipelineCleanedUp =
-      !isFetchingPipelines && hasPipelinesChanged && !state.pipelines;
-
-    if (hasPipelineCleanedUp) fetchPipelines();
-  }, [
-    isFetchingPipelines,
-    state.pipelines,
-    hasPipelinesChanged,
-    fetchPipelines,
-  ]);
-
-  React.useEffect(() => {
-    if (!isFetchingPipelines && !error && pipelines) {
-      dispatch({ type: "LOAD_PIPELINES", payload: pipelines });
-    }
-  }, [dispatch, pipelines, isFetchingPipelines, error]);
-
   return (
-    <ProjectsContext.Provider
-      value={{
-        state,
-        dispatch,
-        fetchPipelines,
-      }}
-    >
+    <ProjectsContext.Provider value={{ state, dispatch }}>
       {children}
     </ProjectsContext.Provider>
   );
